# docs/swagger.yaml
basePath: /
definitions:
  internal_generated_api_post.CreatePostRequest:
    properties:
      body:
        example: This is the post content
        type: string
      title:
        example: My Post Title
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Civil Registry API server with auto-generated documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Civil Registry API
  version: "1.0"
paths:
  /api/post/:
    get:
      consumes:
      - application/json
      description: Retrieve all post records
      produces:
      - application/json
      responses:
        "200":
          description: List of posts
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all posts
      tags:
      - post
    post:
      consumes:
      - application/json
      description: Create a new post record
      parameters:
      - description: post data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_generated_api_post.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created post
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create post
      tags:
      - post
  /api/post/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific post by its ID
      parameters:
      - description: post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: post found
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: post not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get post by ID
      tags:
      - post
  /api/post/health:
    get:
      description: Check if the post service is healthy
      produces:
      - application/json
      responses:
        "200":
          description: Service is healthy
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service is unhealthy
          schema:
            additionalProperties: true
            type: object
      summary: Health check
      tags:
      - post
schemes:
- http
- https
swagger: "2.0"